+++ test/runtests.jl
@@ -35,6 +35,8 @@ start_time = time()
     @memory_profile
     include("accuracy.jl")
     @memory_profile
+    include("strict_mode.jl")
+    @memory_profile
 
     GC.gc()
     rm(tempvidpath, force = true)